requirements.txt
streamlit==1.28.1
opencv-python-headless==4.8.1.78
streamlit-webrtc==0.47.0
numpy==1.24.3
pillow==10.0.0
av==10.0.0
ultralytics==8.0.202
torch==2.0.1
torchvision==0.15.2